requirements.txt
# fastapi uvicorn sqlalchemy psycopg2-binary python-jose[cryptography] passlib[bcrypt] python-multipart

fastapi
uvicorn
sqlalchemy
sqlalchemy-utils
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart